Config API: minor improvements and refactoring.
* When we `set-current-project` it's checked whether valid
  project name is provided;
* Refactoring.
